+++ __tests__/unit/format_query_params.test.ts
@@ -19,7 +19,11 @@ function convertDateToTimezone(date: Date, tz: string) {
 
 describe('formatQueryParams', () => {
   it('formats null', () => {
-    expect(formatQueryParams(null)).toBe('NULL')
+    expect(formatQueryParams(null)).toBe('\\N')
+  })
+
+  it('formats undefined', () => {
+    expect(formatQueryParams(undefined)).toBe('\\N')
   })
 
   it('formats boolean', () => {
@@ -114,14 +118,5 @@ describe('formatQueryParams', () => {
         params: { refs: [44] },
       })
     ).toBe("{'name':'custom','id':42,'params':{'refs':[44]}}")
-  })
-
-  it('throws on unsupported values', () => {
-    expect(() => formatQueryParams(undefined)).toThrowError(
-      'Unsupported value in query parameters: [undefined].'
-    )
-    expect(() => formatQueryParams(undefined)).toThrowError(
-      'Unsupported value in query parameters: [undefined].'
-    )
   })
 })